# splitkauf.openapi.yaml
openapi: "3.0.3"
info:
  title: Splitkauf API
  version: "0.1.0"
  description: |
    Splitkauf API.

  license:
    name: TODO

servers:
  - url: /api/v1

tags:
  - name: system
    description: Service health and operational endpoints.
  - name: identity
    description: The authenticated user.
  - name: lists
    description: Shopping lists and their items.

paths:
  /health:
    get:
      operationId: getHealth
      summary: Service health
      description: |
        Reports overall service health and the status of downstream
        dependencies. Always returns HTTP 200; inspect the payload for status.
      tags:
        - system
      responses:
        "200":
          description: Health status of the service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        default:
          $ref: "#/components/responses/Problem"

  /me:
    get:
      operationId: getMe
      summary: Get the authenticated user
      description: |
        Returns the user injected into the request. In M1 this is always the
        hardcoded dev user (US-A.1); it is replaced by real authentication in M2.
      tags:
        - identity
      responses:
        "200":
          description: The authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"

  /lists:
    get:
      operationId: listLists
      summary: List all lists
      description: |
        Returns every shopping list, each with a summary of its open and checked
        item counts. All lists are visible to every member (no per-user ownership).
      tags:
        - lists
      responses:
        "200":
          description: All lists.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/List"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createList
      summary: Create a list
      description: Creates a new shopping list with the given name.
      tags:
        - lists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListRequest"
      responses:
        "201":
          description: The created list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        default:
          $ref: "#/components/responses/Problem"

  /lists/{listId}:
    parameters:
      - $ref: "#/components/parameters/ListId"
    get:
      operationId: getList
      summary: Get a list with its items
      description: Returns a single list together with all of its items.
      tags:
        - lists
      responses:
        "200":
          description: The list and its items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWithItems"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: renameList
      summary: Rename a list
      description: Renames an existing list.
      tags:
        - lists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameListRequest"
      responses:
        "200":
          description: The renamed list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteList
      summary: Delete a list
      description: Deletes a list and all of its items.
      tags:
        - lists
      responses:
        "204":
          description: The list was deleted.
        default:
          $ref: "#/components/responses/Problem"

  /lists/{listId}/copy:
    parameters:
      - $ref: "#/components/parameters/ListId"
    post:
      operationId: copyList
      summary: Copy a list
      description: |
        Creates a new list containing every non-deleted item of the source list
        (open and checked alike), each reset to unchecked. Name, quantity, unit,
        and note are preserved, as is the source's display order. The request
        body is optional: when no name is supplied the copy is named
        "«Original name» (copy)", trimmed to fit the 200-character name limit.
      tags:
        - lists
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyListRequest"
      responses:
        "201":
          description: The created copy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        default:
          $ref: "#/components/responses/Problem"

  /lists/{listId}/items:
    parameters:
      - $ref: "#/components/parameters/ListId"
    post:
      operationId: addItem
      summary: Add an item to a list
      description: |
        Adds an item to a list. The name is required; quantity defaults to 1;
        the note is optional.
      tags:
        - lists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddItemRequest"
      responses:
        "201":
          description: The created item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        default:
          $ref: "#/components/responses/Problem"

  /lists/{listId}/items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/ListId"
      - $ref: "#/components/parameters/ItemId"
    patch:
      operationId: updateItem
      summary: Update an item
      description: |
        Updates an item's name, quantity, or note. Only the fields present in
        the request are changed (last-write-wins for concurrent edits).
      tags:
        - lists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: The updated item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteItem
      summary: Remove an item
      description: Removes an item from a list.
      tags:
        - lists
      responses:
        "204":
          description: The item was removed.
        default:
          $ref: "#/components/responses/Problem"

  /lists/{listId}/items/{itemId}/check:
    parameters:
      - $ref: "#/components/parameters/ListId"
      - $ref: "#/components/parameters/ItemId"
    post:
      operationId: checkItem
      summary: Check off an item
      description: |
        Marks an item as checked (in the cart). Idempotent: checking an already
        checked item is a no-op that returns the current state.
      tags:
        - lists
      responses:
        "200":
          description: The checked item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        default:
          $ref: "#/components/responses/Problem"

  /lists/{listId}/items/{itemId}/uncheck:
    parameters:
      - $ref: "#/components/parameters/ListId"
      - $ref: "#/components/parameters/ItemId"
    post:
      operationId: uncheckItem
      summary: Uncheck an item
      description: |
        Returns a checked item to the open list. Idempotent: unchecking an
        already open item is a no-op that returns the current state.
      tags:
        - lists
      responses:
        "200":
          description: The unchecked item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        default:
          $ref: "#/components/responses/Problem"

  /lists/{listId}/items/{itemId}/restore:
    parameters:
      - $ref: "#/components/parameters/ListId"
      - $ref: "#/components/parameters/ItemId"
    post:
      operationId: restoreItem
      summary: Restore a deleted item
      description: |
        Restores a soft-deleted item to the list. Idempotent: restoring an item
        that is not deleted is a no-op that returns the current state.
      tags:
        - lists
      responses:
        "200":
          description: The restored item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        default:
          $ref: "#/components/responses/Problem"

components:
  parameters:
    ListId:
      name: listId
      in: path
      required: true
      description: The list's unique identifier.
      schema:
        type: string
        format: uuid
    ItemId:
      name: itemId
      in: path
      required: true
      description: The item's unique identifier.
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: |
        An error occurred. The body is an RFC 9457 Problem Details object with
        Content-Type application/problem+json. Every "type" is a path-absolute
        URI ("/problems/{slug}") that resolves to a human-readable explanation
        page served by the API.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
    Unit:
      type: string
      default: amount
      description: |
        The unit an item's quantity is measured in. A curated German/European
        grocery set. "amount" is the default and is rendered bare (Stück).
        This enum is the single source of truth mirrored by the domain
        (lists.Units) and the items.unit CHECK constraint.
      enum:
        - amount
        - g
        - kg
        - ml
        - l
        - pack
        - bottle
        - can
        - jar
        - cup
        - bunch
        - bag
    ProblemDetail:
      type: object
      description: |
        RFC 9457 (Problem Details for HTTP APIs) error object. The standard
        members are optional; additional extension members may be present.
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri-reference
          description: |
            A URI reference identifying the problem type. Resolves to a
            human-readable explanation page (e.g. "/problems/not-found").
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: The HTTP status code for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.
        errors:
          type: array
          description: Field-level validation errors, when applicable.
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      description: A single field-level validation error.
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable description of the field error.
        pointer:
          type: string
          description: |
            A JSON Pointer (RFC 6901) identifying the offending field within
            the request body.
    HealthStatus:
      type: object
      required:
        - status
        - checks
      properties:
        status:
          type: string
          description: Overall service status (e.g. "ok" or "degraded").
        checks:
          $ref: "#/components/schemas/HealthChecks"
    HealthChecks:
      type: object
      required:
        - database
      properties:
        database:
          type: string
          description: |
            Status of the database dependency ("ok" when reachable, "error"
            when unreachable or not configured).
    User:
      type: object
      description: An authenticated user.
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
          description: The user's unique identifier.
        name:
          type: string
          description: The user's display name.
        email:
          type: string
          format: email
          nullable: true
          description: The user's email address, if known.
    List:
      type: object
      description: A shopping list with a summary of its item counts.
      required:
        - id
        - name
        - openItemCount
        - checkedItemCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: The list's unique identifier.
        name:
          type: string
          description: The list's name.
        openItemCount:
          type: integer
          format: int32
          description: The number of unchecked (open) items on the list.
        checkedItemCount:
          type: integer
          format: int32
          description: The number of checked items on the list.
        createdAt:
          type: string
          format: date-time
          description: When the list was created.
        updatedAt:
          type: string
          format: date-time
          description: When the list was last modified.
    ListWithItems:
      description: A list together with all of its items.
      allOf:
        - $ref: "#/components/schemas/List"
        - type: object
          required:
            - items
          properties:
            items:
              type: array
              description: All items on the list, open and checked.
              items:
                $ref: "#/components/schemas/Item"
    Item:
      type: object
      description: A single item on a shopping list.
      required:
        - id
        - listId
        - name
        - quantity
        - unit
        - checked
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: The item's unique identifier.
        listId:
          type: string
          format: uuid
          description: The identifier of the list this item belongs to.
        name:
          type: string
          description: What to buy.
        quantity:
          type: integer
          format: int32
          default: 1
          minimum: 1
          description: How many to buy. Defaults to 1.
        unit:
          $ref: "#/components/schemas/Unit"
        note:
          type: string
          nullable: true
          description: An optional free-text note (e.g. brand or size).
        checked:
          type: boolean
          description: Whether the item has been checked off (in the cart).
        checkedAt:
          type: string
          format: date-time
          nullable: true
          description: When the item was last checked off, if it is checked.
        createdAt:
          type: string
          format: date-time
          description: When the item was created.
        updatedAt:
          type: string
          format: date-time
          description: When the item was last modified.
    CreateListRequest:
      type: object
      description: Request body for creating a list.
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: The name for the new list.
    RenameListRequest:
      type: object
      description: Request body for renaming a list.
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: The new name for the list.
    CopyListRequest:
      type: object
      description: |
        Optional request body for copying a list. When omitted (or when `name`
        is absent) the server names the copy "«Original name» (copy)".
      properties:
        name:
          type: string
          minLength: 1
          description: The name for the copy.
    AddItemRequest:
      type: object
      description: Request body for adding an item to a list.
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: What to buy.
        quantity:
          type: integer
          format: int32
          default: 1
          minimum: 1
          description: How many to buy. Defaults to 1.
        unit:
          $ref: "#/components/schemas/Unit"
        note:
          type: string
          nullable: true
          description: An optional free-text note.
        checked:
          type: boolean
          default: false
          description: |
            Whether the item is checked off at creation. Used when an offline
            check folds into a still-queued create; defaults to false.
    UpdateItemRequest:
      type: object
      description: |
        Request body for updating an item. Only the fields present are changed.
      properties:
        name:
          type: string
          minLength: 1
          description: A new name for the item.
        quantity:
          type: integer
          format: int32
          minimum: 1
          description: A new quantity for the item.
        unit:
          $ref: "#/components/schemas/Unit"
        note:
          type: string
          nullable: true
          description: A new note for the item (null clears it).